Fourth and Final version of the Encrypt Word assignment
This is the final iteration of the EncryptWord assignment. In this assignment I have included a large number of test words, and this iteration is by far the most complete and functional of all of the versions. The driver works by asking the user to input a desired length for a word for them to guess, then encrypts that word using a randomly chosen class from amongst the three different word-encryption classes, and gives the user 3 chances to guess the encrypted word. The program keeps track of guesses. 
The list of available words has been censored to remove fowl/inappropriate language, and some of the encryptions can be fairly challenging to guess.